apps/api/pyproject.toml
[project]
name = "api"
version = "0.0.13"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
 "bcrypt<4.0.0",
 "beautifulsoup4>=4.14.2",
 "boto3>=1.34.0",
 "bson>=0.5.10",
 "celery>=5.5.3",
 "faker>=37.11.0",
 "fastapi>=0.119.0",
 "markdown>=3.7",
 "passlib>=1.7.4",
 "psycopg2-binary>=2.9.11",
 "pydantic-settings>=2.11.0",
 "pyjwt>=2.10.1",
 "python-multipart>=0.0.20",
 "python-opengraph>=0.0.2",
 "redis>=6.4.0",
 "requests>=2.32.5",
 "scalar-fastapi>=1.4.3",
 "sqlmodel>=0.0.27",
 "uvicorn>=0.38.0",
 "alembic>=1.13.0",
 "python-dotenv>=1.1.1",
 "httpx-oauth>=0.15.0",
 "resend>=2.19.0",
 "loguru>=0.7.3",
 "websockets>=12.0",
]

[dependency-groups]
dev = [ "ruff>=0.14.1", "ty>=0.0.1a23", "pytest>=8.0.0", "httpx>=0.25.0", "pytest-asyncio>=0.21.0", "python-semantic-release>=9.0.0", "toml>=0.10.2",]
